requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pymongo==4.13.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.4
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.4.0
user-agents==2.2.0
Werkzeug==3.1.3
